Combine koin modules to remove clutter #127
