+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/update/batch/ChunkBatches.kt
@@ -2,7 +2,6 @@ package world.gregs.voidps.engine.client.update.batch
 
 import it.unimi.dsi.fastutil.objects.ObjectLinkedOpenHashSet
 import kotlinx.io.pool.DefaultPool
-import org.koin.dsl.module
 import world.gregs.voidps.engine.client.update.view.Viewport
 import world.gregs.voidps.engine.entity.MAX_PLAYERS
 import world.gregs.voidps.engine.entity.character.player.Player
@@ -119,8 +118,4 @@ class ChunkBatches(
         private const val EXPECTED_PLAYERS_PER_CHUNK = 8
         private const val EXPECTED_UPDATES = EXPECTED_PLAYERS_PER_CHUNK * 2
     }
-}
-
-val batchedChunkModule = module {
-    single { ChunkBatches() }
 }
