+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/World.kt
@@ -5,6 +5,9 @@ import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.engine.map.Tile
 import world.gregs.voidps.engine.utility.get
 
+const val MAX_PLAYERS = 0x800 // 2048
+const val MAX_NPCS = 0x8000 // 32768
+
 object World : Entity {
     override var tile = Tile.EMPTY
     override val size: Size = Size.ONE
