+++ engine/src/main/kotlin/world/gregs/voidps/engine/event/EventHandlerStore.kt
@@ -1,6 +1,5 @@
 package world.gregs.voidps.engine.event
 
-import org.koin.dsl.module
 import world.gregs.voidps.engine.entity.Entity
 import world.gregs.voidps.engine.entity.World
 import world.gregs.voidps.engine.entity.character.Character
@@ -14,10 +13,6 @@ import world.gregs.voidps.engine.entity.proj.Projectile
 import world.gregs.voidps.engine.entity.sound.AreaSound
 import world.gregs.voidps.engine.utility.get
 import kotlin.reflect.KClass
-
-val eventModule = module {
-    single { EventHandlerStore() }
-}
 
 /**
  * Handles the storage and delivery of global [EventHandler]'s
