+++ engine/src/main/kotlin/world/gregs/voidps/engine/map/area/Areas.kt
@@ -1,14 +1,9 @@
 package world.gregs.voidps.engine.map.area
 
-import org.koin.dsl.module
 import world.gregs.voidps.engine.data.FileStorage
 import world.gregs.voidps.engine.timedLoad
 import world.gregs.voidps.engine.utility.get
 import world.gregs.voidps.engine.utility.getProperty
-
-val areasModule = module {
-    single(createdAtStart = true) { Areas().load() }
-}
 
 class Areas {
