+++ engine/src/main/kotlin/world/gregs/voidps/engine/map/region/Xteas.kt
@@ -1,7 +1,5 @@
 package world.gregs.voidps.engine.map.region
 
-import org.koin.dsl.module
-
 data class Xteas(
     val delegate: MutableMap<Int, IntArray> = mutableMapOf()
 ) : Map<Int, IntArray> by delegate {
@@ -10,12 +8,4 @@ data class Xteas(
         return this[region.id]
     }
 
-}
-
-val xteaModule = module {
-    single(createdAtStart = true) {
-        Xteas(mutableMapOf()).apply {
-            XteaLoader().load(this, getProperty("xteaPath"), getPropertyOrNull("xteaJsonKey"), getPropertyOrNull("xteaJsonValue"))
-        }
-    }
 }
