+++ tools/src/main/kotlin/world/gregs/voidps/tools/detail/AnimationNames.kt
@@ -1,6 +1,8 @@
 package world.gregs.voidps.tools.detail
 
 import org.koin.core.context.startKoin
+import org.koin.core.qualifier.named
+import org.koin.dsl.module
 import org.koin.fileProperties
 import world.gregs.voidps.cache.Cache
 import world.gregs.voidps.cache.config.decoder.RenderAnimationDecoder
@@ -10,7 +12,6 @@ import world.gregs.voidps.cache.definition.decoder.NPCDecoder
 import world.gregs.voidps.engine.client.cacheDefinitionModule
 import world.gregs.voidps.engine.client.cacheModule
 import world.gregs.voidps.engine.data.FileStorage
-import world.gregs.voidps.engine.data.fileStorageModule
 import world.gregs.voidps.engine.entity.definition.DefinitionsDecoder.Companion.toIdentifier
 
 /**
@@ -25,7 +26,10 @@ object AnimationNames {
     fun main(args: Array<String>) {
         val koin = startKoin {
             fileProperties("/tool.properties")
-            modules(cacheModule, cacheDefinitionModule, fileStorageModule)
+            modules(cacheModule, cacheDefinitionModule, module {
+                single { FileStorage() }
+                single(named("jsonStorage")) { FileStorage(json = true) }
+            })
         }.koin
         val cache: Cache = koin.get()
         val storage: FileStorage = koin.get()
